Remove setup steps that are PE specific
Previously we put all of the setup steps for all of Puppetlabs projects
in its test harness. As projects change the maintenance cost should be
shifted to the developers/projects that are changing. They are the best
to understand their particular projects setup/dependencies. We have
moved all of the PE setup tests into a separate repository but the
harness will still try to load these outdated setup steps. This cleans
up the duplicate and/or conflicting setup steps that are better specified in
the PE projects test suite.