Specify if we are doing a PE run via env var
Previously we determined if we were running a PE suite based on what
setup option was chosen. As we move the setup of test suite out of the
harness the flags that would have told us whether or not we were running
a PE suite are no longer of use. Ideally we would allow the project
being tested to override the host defaults for values, but this that
would be a larger change than is currently scoped. In the mean time this
patch allows setting an environment variable to determine wether or not
this is a PE or FOSS suite run.