Ensure the consoleport is always a string.
Previously, the consoleport would be a string if set by the
environment or a config file, but an integer if left at its
default. This changes the default to a string.

I'm changing the default to a string rather than converting the
env/cfg value to an integer because we have historically set the
consoleport in our config files, so this way it is less likely to
break things.