(maint) Do not fall back to the user-specified VM name when doing vcloud cleanup
The vcloud provisioning will never create a VM with the user-specified
name, so falling back to it was pointless, and could cause the harness
to delete VMs not associated with the current test run, or worse,
templates.

As a special bonus, if any hosts don't have vmhostname set, we will
warn that they were likely not provisioned correctly in the first
place.