(maint) Correct solaris 11/10 package install/uninstall
Solaris has different package handling between 11 and 10.  We were
correct on install for Solaris 11, but uninstall was Solaris 10.

Now differentiate between the two with the correct package handlers for
11 and 10 respectively.

This also corrects a bug checking for existing packages in Solaris 10.
The which command returns an exit code of 0, even if the command is not
found.  So we're testing stdout explicitly to see if a path to command
was returned.