QE-364 Clean up line endings
When using `Beaker::Command` helpers such as `facter()` or even `on()`
the resulting output is often a single line and of interest for its
value. The newline returned at the end of the line is not relevant to
interests and ends up getting `.chomp` called on it in almost every
case.

This change chomps the final newline off of stdout and stderr output so
that tests can be more simply written without having to `.chomp`