Relax regex for windows group names
Windows::Group.group_list was only recording group names matching
/[\w ]/, but windows groups have a much larger set of valid characters
(http://technet.microsoft.com/en-us/library/bb726984.aspx).  This was
causing a failure on a Windows 2012 template which had a group with a
'-' in it.

Instead of trying to specify what is valid, accept whatever the wmic
group indicates is a group, since it should know.