Install agents using the frictionless agent install
This change adds logic for installing using the new curlbash style of
agent install. In order to support testing both the old and new styles
of install, we add a new agent role of "frictionless", which indicates
that the node should be installed in the new style. This simply causes
it not to download the PE tarball or generate answer files, and to curl
and run a script from the master (based on the platform tag).