(QA-747) Log the amount of time each command takes to execute
Previously, there was no way to tell how long each beaker command took
to execute on a remote VM. I would like to know this information to
track down a timing issue in acceptance testing on Windows.

This commit modifies the host object to log timing information when
beaker is run in debug mode. It uses the same logging prefix as is used
when displaying the executed command, e.g.

    a4e5hyo2m0ooom1 (master) $  cp /etc/puppet/puppet.conf ...
    a4e5hyo2m0ooom1 (master) executed in 0.01 seconds