(QA-753) Don't write or read the PE tarball more than we have to
The PE tarball is big and our storage isn't always happy about
that. By streaming the tarball unpack, we can avoid a write and then a
read of the 800MB tar. On a matrix of 200 nodes, that comes up to 20
million 4K blocks written then read, for no real gain.