(QA-773) Support installing any frictionless agent on any master
Previously, beaker only supported installing a frictionless agent if
it had the same platform as the master. This code allows heterogenous
deployments by properly classifying the master for the platform of
each frictionless node in the configuration.

This is (once again) kind of a quick-and-dirty solution. It'll get
cleaned up as part of QA-769, wherein we refactor the mess known as
the PE installation code.