Patch RS_DESTROY behavior
QA-723's implementation moved the logic for `@options[:preserve_hosts]`
out of the `Beaker::NetworkManager#cleanup` method to
`Beaker::CLI#execute!` and beaker-rspec doesn't use the CLI, so the
logic would have to be duplicated here, or beaker would need a patch to
move it back.

It doesn't really make sense for the cleanup method to conditionally
cleanup, and it doesn't make sense for beaker to be inspecting the
success state of RSpec, so beaker-rspec can be changed to work.