add safe_popen3 to Beaker::Hypervisor::Vagrant
When running beaker under bundler the environment setup leaks out into sub
shells.  This is specifically a problem for Vagrant on Gentoo which is not a
vendored install. The safe_popen3 method is a wrapper around Open3.popen3 that
detects if beaker is running and cleans up the environment as necessary.