(QA-845) Allow PE upgrade version to be specified in the config
Previously, when doing an upgrade beaker could only get a version by
reading the LATEST file. This makes it difficult to track exactly
which version is going to be tested in a given test run. By allowing
the version to be specified in the config file, we can make sure
beaker installs precisely the PE version that we care about for
testing.