(QENG-64) Fix ordering issue when deploying the frictionless agent
Previously, the puppet agent was not stopped on the master until all
nodes were deployed. This could cause a situation where the background
agent would be running when we attempt to deploy the frictionless
repo. In this situation, the agent run we kick off to deploy the repo
would fail. This causes the entire beaker process to bail.

We now stop the background agent immediately, preventing this
condition from occuring.