Ensure IS_PE only respects different values
Previously if we set the IS_PE or BEAKER_IS_PE environment variable at
all we would automatically make the run a PE run. This caused
non-obvious behavior for users that set IS_PE=false.

This changes the searching of the environment to only set the :type
option if [BEAKER_]IS_PE=yes|true.

Signed-off-by: Justin Stoller <justin@puppetlabs.com>