(QENG-842) Always wait for puppetmaster to be ready when bouncing
Previously, we would only wait for the puppetmaster to be started if a
host config file explicitly requested it. Our init scripts actually
require this waiting pretty much across the board now, so trying to
only do this when it's needed is silly. And even when it's not needed
we only curl once - it's not a ton of overhead.