(maint) Improve `InstallUtils#install_puppet()`
Use `#check_for_command` instead of `#check_for_package` where appropriate
in `#install_puppet`

Previously there was no `Host#check_for_command` method, but
`Host#check_for_package` did a `which` (it effectively was checking for
the command). Now that we've separated out that behavior we need to use
the ones we mean in `InstallUtils#install_puppet()`. This is important
because when we fall back to a gem install of Puppet we ensure there is
a command `gem` available from (often) the package `rubygems`.