(QENG-1000) Ensure that add-el-extras works on all rhel platforms
The --add-el-extras options was only working for platform strings of the form
el-(5,6), which meant that you couldn't have beaker add the el repo on a
'centos' platform for instance.

Change makes more use of the Beaker::Platform attributes and checks
other rhel variants (centos, scientific, and the 'redhat' variant is
another platform label used for rhel I believe).