(gh-471) Adding support to vagrant for box_version and box_check_update.
If using versioned boxes from vagrant cloud and a new update is
published, by default vagrant will attempt to update that box to
the latest version. This may be undesired behaviour. Like any other
versioned object you way wish to fix that initial download to a
fixed version or a version range.