(gh-426) Add prebuild packages to Dockerfile
Without this patch every time the specs run, prebuilt packages are
installed to the new docker image. Even when preserving the image.
To increase the speed of the test suite we install those packages when
creating the Docker image. The `HostPrebuiltSteps` checks whether those
are installed and does not do anything directly starting to execute the
specs inside the image.