(gh-469) Adding a new wrapper for the powershell.exe command
There are several places in the codebase where this will be used as
powershell is the most common scipting environment on windows.

The aim of this new addtion is to wrap a few of the complexities of
executing powershell commands, such as discovering the powershell.exe
location and passing the correct command line paramaters.