Add VPC support to the aws_sdk hypervisor
This adds support for launching instances into an AWS VPC.

This adds two new configuration options:

- subnet_id
- vpc_id

That can be added to the HOSTS section of your beaker configuration. These
options are both mandatory for this feature to work.

The code has been basically modified to deal with the VPC object now,
if those configuration items are provided.

Signed-off-by: Ken Barber <ken@bob.sh>