Allow frictionlesss install to pass params
Previously, the frictionless agent install did not have a
way to specify puppet config settings at install time. In 3.4 the
ability was added to pass in additional settings in the form of
`section:key=value` and have it call `puppet config set` with those
options before starting puppet. This allows an agent to dynamically
configure it's certname, server or any other setting.

Conflicts:
	lib/beaker/dsl/install_utils.rb