Bump aws-sdk to v1.57.0 for eu-central-1 support
Prior to this commit, we pinned aws-sdk to v1.42.0. Today, AWS brought
the eu-central-1 datacenter online. When Beaker attempts to run its
zombie killer script(s), the AWS SDK doesn't know how to handle this
datacenter. To combat the issue, AWS released v1.57.0 of the aws-sdk gem
today to coincide with this datacenter being brought online.

This commit modifies Beaker's Gemspec to use aws-sdk v1.57.0 so that it
can properly use the eu-central-1 datacenter.