(QENG-1400) Allowed yield blocks to get at actual command output, and returned unix style passwd user output
Mac's dscl output isn't formatted as the /etc/passwd expected by callers.  In order to go back to this, I changed the yield to use the
information after I formatted it.  However, this fails because the callers are relying on the parameter given to them having a .stdout
call and other things like it, so I've gone back to that old way, and allowed the unix formatting to still be used in the return