(MAINT) pinned Yard version
Yard moved to 0.8.7.6 over the weekend, which broke our tests. This is an
issue because it takes time to figure out what happened when this upgrades
from under us when we're not aware, and they end up failing.  Pinning this
makes the problem explicit when we take the explicit action of upgrading
the library.  Pinned to 0.8.7.4.