(QENG-1454) Allow user to specify ports to open on ec2
Previous to this commit, the list of ports that beaker should open for
ec2 was hard coded and incomplete. For example it was missing port
'5432' for the database node, which is the port postgresql uses.
This commit adds the ability for a user to specify additional ports to open
in their `hosts.cfg` file. By allowing this, it should provide some
flexibility to the hard coded nature of the ports, and allow beaker to
just provide the defaults that are currently known.