(QENG-1430) Added fail_on_error option to host.exec
Before, if an unacceptable exit code was returned, there was only one result: the
test would error out.  QA wants to be able to specify that as a failure in their
tests.  I've added this functionality in that if the fail_on_error option is set,
we fail rather than error out.