(QENG-1430) Added accept_all_exit_codes option to host exec calls
QA was having issues with errors occuring when they'd like to have failures, since the specific
error codes are forseen and specific failure cases.  In order to help them with this use case,
I've added the accept_all_exit_codes option, which will not error on unacceptable exit codes,
defering to the test above to deal with the command's exit code as a failure or otherwise.