(gh-585) Add support for vagrant 1.7.x
As of version 1.7.x, vagrant will replace the default insecure keypair with a
randomly generated keypair on first 'vagrant up'. As a result the generated
ssh configs have changed and contain per-host private keys. Beaker was
breaking the path to these new private key files due to a substitution which
wasn't sufficiently constrained (the path to the private key now contains the
hostname, which was being replaced with the host's IP). This commit adds the
necessary constraint.