Add a libvirt hypervisor
This commit adds a libvirt hypervisor that shells out to virsh.
This does not use the libvirt ruby gem. The libvirt ruby gem is
not well maintained and gets us into dependency problems with
nokogiri fast.

Dependency issues with nokogiri is what motivated me to write this
instead of using the existing libvirt-vagrant hypervisor

I'll likely write a follow on blog post about how I build my qcow
images and use this hypervisor.