(BKR-217) aws_sdk: given CONFIG subnet_ids, try each in turn
This should help alleviate the EC2 InsufficientInstanceCapacity.

The new policy will only be applied to a host if there isn't a CONFIG
subnet_id and there isn't a host-specific subnet_id.  When it does
apply, try each subnet in subnet_ids once for each relevant host.

The new option can be used like this:

  CONFIG
    vpc_id: foo
    subnet_ids:
      - x
      - y
      - z