(BKR-195) Allow installation of puppet-agent
This allows the installation of puppet-agent via the install_puppet()
helper when passed a version of puppet >= 4.0.0. What version that will
be installed is configurable via the :puppet_agent_version option. This
option is required for installing Windows MSIs.

This is meant to be a short term feature that allows module authors to
try puppet-agent in their testing with minimal changes. Long term a
dedicated method or making AIO the default for install_puppet() may be
the solution.