(BKR-249) Add OpenBSD support
Add install method to prefer installing the OpenBSD-maintained Puppet package
which tends to work better than vanilla gem versions. Also add enough smarts
to be able to do gem installations. The shell is not a login shell therefore
it doesn't pick up the .profile that contains the PKG_PATH variable so pkg_add
doesn't automatically install from the network so add the PKG_PATH to the
.ssh/environment too.

The packaging support handles the case when there are multiple rubies
available. Currently it will install the newest in the list that is older than
2.2.x as that doesn't work with all versions of Puppet. Also, if installing the
package advises to create symlinks to make this package the default, perform
this step automatically. This handles for example when installing ruby-1.9.3
and all the commands have a 19 suffix, etc.