(BKR-242) Add Swarm support to docker run
To distribute the same built image between all swarm slaves we need to
push the image to a private registry.
If the image does not exist yet ( first build ) we tag + push it.
If it does exist we will re-use it.
If we built the image on Node A but then the run command is executed on
node B we will pull the image automatically and run it.

The only downside is that we will always build the image since we need
to know the resulting image ID.
Hopefully this can be improved in the future.