(BKR-311) Allow AWS IP Fallback
When bringing up an instance in a VPC there won't always be a public address.
This allows the AWS hypervisor to fallback and assign an instance's private
ip to host['ip'] for new instances without a public address.

This PR includes tests for the .AwsSdk#populate_dns method to verify this new
behavior persists in the future.