(BKR-301) New test and testing seams
This splits off tests for #load_credentials, as we have untested
behavior there, and the point of this work is to fix a mis-feature
in that area (not catching the hard-to-decipher exception when
an empty fog file is presented).  The main tests all stub out the
fog file loading, so here we introduce a different stub.

This also modifies the `Vmpooler` class to split out a testing seam:
the `#read_fog_file` method can be stubbed now independently of the
`#load_credentials method`, so we can feed in post-YAML output to
the `#load_credentials` under test.

Since `#load_credentials` is called as part of the constructor (a mild
anti-pattern, but lazy loading can be cumbersome), we need to validate
its results via testing post-constructed object state.  Unfortunately,
we can't get the resultant credentials state without piercing the
instance-variable boundary of the `Vmpooler` object, so we added another
seam: the `attr_reader` for `:credentials`.

While we're there, we added the reader for the remainder of the
options handed into the constructor, as it makes sense to at least
be uniform in the library, even if it is under-tested.