(BKR-195) Simplified install of released puppet-agent
Previously installing the puppet-agent package required writing a
pre-suite to manually install the correct dev repositories and
installing a dev build of puppet-agent.

This is problematic for downstream consumers that do not want to test
puppet-agent or the code in it, but want to test that their code works
with puppet-agent (eg module authors).

This commit creates the DSL method `install_puppet_agent_on()`, an
analog to the `install_puppet_on()` method that allows installing
puppet-agent.

It also allows the `install_puppet()` method to install puppet-agent if
given a version of Puppet greater than 4.0.0 (by delegating to
`install_puppet_agent_on()`).

Errata:
 * Follows the pattern of naming helpers `*_on` and
   normalizes the helpers it touches to follow the same pattern.
 * Updates the MSI and DMG methods to allow for latest builds
   (prior to this they required explicit versions and so were rarely
   used even by our internal teams).
 * Also updates legacy puppet MSI and DMG install helpers to
   delegate to newer install puppet-agent helpers using the same
   criteria as used in `install_puppet_on()` (despite being marked as
   api private they are used internally).
 * Uses newer `install_puppetlabs_release_repo_on` and
   `Host#install_package` and updates old helpers to do the same