(BKR-53) with_puppet_running_on can not restart after yielding
Before, with_puppet_running_on bounced the puppet service before and after runs.
This was causing us a lot of pain as the last restart proved unnecessary for the
most part, and added 30 seconds to each test run on avg in our CI.

We've decided to improve this by adding the ability to remove the second bounce.
Mosts tests use `with_puppet_running_on`, so they'll be setup accordingly before
yielding, so that second one is unnecessary.