(BKR-339) Remove the target module directory on host during copy_module_to
Without this patch, the copy_module_to function will not work correctly to copy updated module files into the target directory on the hosts. Instead accpetance test spec files can be updated, but not module sources. Additionally the module copy will fail on the second non-provisioning run.

The patch simply executes a delete of the target directory on the host, ensuring it does not exist prior to the mv command.