(BKR-366) merge up install from git changes from deployer and puppet
This change merges changes to install_from_git that were overloaded into
puppet and other repos' acceptance libraries.
We also split the cloning portion of install_from_git into its own
method, added options that are passed through to on() and moved a method
to beaker's acceptance library that needs to be shared across tests
(install_packages_on).
This addition of an acceptance library for beaker may require changes to
the jenkins CI configuration, to add the library to the test-runner's
load-path.
The merge-up also brought in helpers for install_from_git,
build_git_url() and lookup_in_env().
This change also adds a few more features to build_git_url that were required
for deployer, including giturl protocol.