(maint) Allow PE puppet-agent installs on custom roles
Previously we installed puppet-agent on a host if there was the agent
role and no other role assigned to it. This prevents assigning a
puppet-agent node the "default" role to be used as the default test
target, or allow any other custom roles.

To allow assigning the "default" or custom roles to a puppet-agent node
we list the non-puppet-agnet roles we know how to generate installation answers
for and use that list to determine what nodes are puppet-agent or not.