(BKR-401) fix race in OpenStack volume deletion
There is a delay between a volume being detached from a virtual machine and
when it is ready to be deleted.  The code as it stands is likely to hit a 400
error as the volume is still in the 'detaching' state.  Add in a wait_for
statement to allow the volume to get back into the 'available' state before
deleting