(BKR-411) Set host.ip for Fusion hypervisor
This commit updates the Fusion hypervisor to set the host[:ip]
value based on the IP address obtained from querying the VM data
using the fission gem.

This removes the need for the user to have setup dns resolution to
the `host.name` prior to running beaker.

This commit also amends the rspec mock for fission to respond to
the `network_info` method. See the following link to the method
in the fission gem

https://github.com/thbishop/fission/blob/master/lib/fission/vm.rb#L201