(BKR-411) Fusion hypervisor: use host.name when possible, add dns
This commit ensures that the host[:ip] is not set if the host.name
can be connected to.

In the event that host[:ip] is set, It also adds methods for adding
the ip and hostname to all of the hosts to ensure internal name
resolution between them.

Spec tests for the above behavior are also added.