(maint) use type over cat in pure Windows
If the host is not a cygwin environment, prefer `type` to produce the
contents of the log file over `cat`. `type` is built-in, where `cat` is
only available if someone has put *nix tools on the path, whether
through a git install (MinGW tools), installing GnuWin utils or some
other method.