(maint) Change assert Method; Workaround debian Check for Package
assert_equal(true,...) is the same as just assert(...). Change all
of those matching calls to just assert.

beaker can incorrectly report a package as being installed when in fact it is not. This is specific to debian systems. As a work around to the above, debian systems will use check_for_command to determine whether the specific package is installed.