(BKR-358) Add acceptance tests for scp_to
This is a work in progress commit. I want to push up state because it turns out there is an
ordering dependency in these tests; if scp_to fails due to the remote path not existing, it
throws a RuntimeError, which appears to disable further SSH connections to the remote host.
Subsequent tests will just hang. Probably a bug in beaker.